Moved ListManager logic to SimpleAdapter.kt
